printf pronta

+++ myprintf/printf-bonus/ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: bgomes-b <bgomes-b@students.42.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/03 12:49:10 by bgomes-b          #+#    #+#             */
-/*   Updated: 2026/02/12 22:13:26 by bgomes-b         ###   ########.fr       */
+/*   Updated: 2026/02/13 05:03:20 by bgomes-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,10 +44,10 @@ static char	*ft_strdup(const char *s)
 	return (str);
 }
 
-static int	numlen(int n)
+static int	numlen(long n)
 {
-	int	len;
-	int	num;
+	int		len;
+	long	num;
 
 	len = 0;
 	num = n;
@@ -83,10 +83,10 @@ static void	rstring(char *s)
 	}
 }
 
-char	*ft_itoa(int n)
+char	*ft_itoa(long n)
 {
-	int		i;
-	int		sign;
+	long	i;
+	long	sign;
 	char	*s;
 
 	if (n == -2147483648)
